pyproject.toml
[tool.poetry]
name = "pram"
version = "0.0.1"
description = ""
authors = ["Rory Byrne <rory@rory.bio>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
pydantic = "^2.5.3"
torch = "^2.1.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
pylint = "^3.0.3"
isort = "^5.13.2"
flake8 = "^7.0.0"
pyright = "^1.1.345"
flake8-pyproject = "^1.2.3"
tox = "^4.12.0"
coverage = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

##############
### CONFIG ###
##############

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint]
max-line-length = 120
# extension-pkg-whitelist = "pydantic"
good-names = ["fp"]
disable = ["logging-fstring-interpolation"]

[tool.flake8]
max-line-length = 120
extend-ignore = "E266"
disable = ["logging-fstring-interpolation"]

[tool.pyright]
include = ["pram", "tests" ]

[tool.pytest]
asyncio_mode = "auto"

[tool.pytest.ini_options]
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
